==> experiments_archive/515/New_PIH/4_label.ipynb <==
import os
import pandas as pd
import shutil

train angle

cropped_rgb_path = './dataset/train/cropped/rgb/'
cropped_depth_path = './dataset/train/cropped/depth/'

angle_rgb_path = './dataset/train/angle/rgb/'
angle_depth_path = './dataset/train/angle/depth'

cropped_train_dir = './dataset/train/cropped/train_data.csv'
angle_train_dir = './dataset/train/angle/train_angle.csv'

cropped_train_data = pd.read_csv(cropped_train_dir)

angle_count = {}
angle_data = []

for index, row in cropped_train_data.iterrows():
    # Extract the angle value from the DataFrame
    angle_value = format(row['Angle [deg]'],".5f")

    # Check if the angle value already exists in the dictionary
    if angle_value not in angle_count:
        angle_count[angle_value] = -1
    
    # Increment the count for the angle value and format it properly
    angle_count[angle_value] += 1
    
    # Construct the new image names with unique suffixes based on the count
    new_name = f"{angle_value}{angle_count[angle_value]}.png"
    
    # Define the original image paths
    original_rgb_name = row['Picture name']
    original_rgb_path = os.path.join(cropped_rgb_path, original_rgb_name)
    original_depth_path = os.path.join(cropped_depth_path, original_rgb_name)
    
    # Calculate the angle range folder based on integer division by 22.5
    angle_folder = str(int(float(angle_value) // 22.5)).zfill(2)
    
    # Define the new image paths in the angle folders
    new_rgb_path = os.path.join(angle_rgb_path + '/' + angle_folder + '/', new_name)
    new_depth_path = os.path.join(angle_depth_path + '/' + angle_folder + '/', new_name)
    
    # Copy the images to the angle folders with the new names
    shutil.copyfile(original_rgb_path, new_rgb_path)
    shutil.copyfile(original_depth_path, new_depth_path)
    
    angle_data.append([new_name, row['Radial distance [mm]']])

# Convert the angle data list to a DataFrame
angle_df = pd.DataFrame(angle_data, columns=['Picture name', 'Radial distance [mm]'])

# Save the DataFrame to a new CSV file
angle_df.to_csv(angle_train_dir, index=False)

train rad

cropped_rgb_path = './dataset/train/cropped/rgb/'
cropped_depth_path = './dataset/train/cropped/depth/'

rad_rgb_path = './dataset/train/rad/rgb/'
rad_depth_path = './dataset/train/rad/depth'

cropped_train_dir = './dataset/train/cropped/train_data.csv'
rad_train_dir = './dataset/train/rad/train_rad.csv'

cropped_train_data = pd.read_csv(cropped_train_dir)

rad_count = {}
rad_data = []

for index, row in cropped_train_data.iterrows():
    # Extract the rad value from the DataFrame
    rad_value = format(row['Radial distance [mm]'],".5f")

    # Check if the rad value already exists in the dictionary
    if rad_value not in rad_count:
        rad_count[rad_value] = -1
    
    # Increment the count for the rad value and format it properly
    rad_count[rad_value] += 1
    
    # Construct the new image names with unique suffixes based on the count
    new_name = f"{rad_value}{rad_count[rad_value]}.png"
    
    # Define the original image paths
    original_rgb_name = row['Picture name']
    original_rgb_path = os.path.join(cropped_rgb_path, original_rgb_name)
    original_depth_path = os.path.join(cropped_depth_path, original_rgb_name)
    
    # Calculate the radial distance range folder based on integer division by 1
    rad_folder = str(int(float(rad_value) // 1))
    
    # Define the new image paths in the rad folders
    new_rgb_path = os.path.join(rad_rgb_path + '/' + rad_folder + '/', new_name)
    new_depth_path = os.path.join(rad_depth_path + '/' + rad_folder + '/', new_name)
    
    # Copy the images to the rad folders with the new names
    shutil.copyfile(original_rgb_path, new_rgb_path)
    shutil.copyfile(original_depth_path, new_depth_path)
    
    rad_data.append([new_name, row['Angle [deg]']])

# Convert the rad data list to a DataFrame
rad_df = pd.DataFrame(rad_data, columns=['Picture name', 'Angle [deg]'])

# Save the DataFrame to a new CSV file
rad_df.to_csv(rad_train_dir, index=False)
